Remove more variables, hardcode them to simplify for now
