+++ src/sound_buffer.h
@@ -1,5 +1,4 @@
 #pragma once
-#pragma comment(lib, "winmm.lib")
 #include <Windows.h>
 
 #include <vector>
@@ -7,6 +6,13 @@
 
 namespace blga {
 
+struct block_proxy
+{
+    WAVEHDR*         header;
+    std::span<short> data;
+
+    void send_to_sound_device(HWAVEOUT device) const;
+};
 
 class audio_buffer
 {
@@ -16,23 +22,6 @@ class audio_buffer
         std::vector<short> data;
     };
 
-    struct block_proxy
-    {
-        WAVEHDR*         header;
-        std::span<short> data;
-
-        void send_to_sound_device(HWAVEOUT device) const
-        {
-            // Prepare block for processing
-			if (header->dwFlags & WHDR_PREPARED) {
-				waveOutUnprepareHeader(device, header, sizeof(WAVEHDR));
-            }
-
-            waveOutPrepareHeader(device, header, sizeof(WAVEHDR));
-			waveOutWrite(device, header, sizeof(WAVEHDR));
-        }
-    };
-
     std::vector<block> d_blocks;
     std::size_t        d_current;
   
@@ -40,29 +29,9 @@ class audio_buffer
     audio_buffer& operator=(const audio_buffer&) = delete;
 
 public:
-    audio_buffer(std::size_t num_blocks, std::size_t block_size)
-        : d_blocks(num_blocks)
-        , d_current(0)
-    {
-        for (auto& block : d_blocks) {
-            block.data.resize(block_size, 0);
-            block.header.dwBufferLength = static_cast<DWORD>(block.data.size() * sizeof(short));
-            block.header.lpData = (LPSTR)block.data.data();
-		}
-    }
+    audio_buffer(std::size_t num_blocks, std::size_t block_size);
 
-    block_proxy next_block()
-    {
-        // Get the next block to use
-        auto& current = d_blocks[d_current];
-        ++d_current;
-        d_current %= d_blocks.size();
-
-        return {
-            .header = &current.header,
-            .data = current.data
-        };
-    }
+    block_proxy next_block();
 };
 
 }
